requirements.txt
pydub==0.25.1
PyYAML==6.0
librosa==0.9.2
numpy==1.24.1
